% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{analyze_fastq}
\alias{analyze_fastq}
\title{Return string \code{"Hello world!"} to R.}
\usage{
analyze_fastq(file_path, min_length, min_avg_qual)
}
\description{
Return string \code{"Hello world!"} to R.
}
