% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{analyze_fastq_r}
\alias{analyze_fastq_r}
\title{produces a dataframe of all the reads, min length and GC content under a certain threshold.}
\usage{
analyze_fastq_r(file_path, min_length, min_avg_qual, min_gc_content)
}
\description{
produces a dataframe of all the reads, min length and GC content under a certain threshold.
}
